add animation to demonstrate dfs add user friendly stack output

+++ Graph/Graph/src/com/gmail/at/ivanehreshi/interfaces/GraphicManipulator.java
@@ -1,5 +1,5 @@
 package com.gmail.at.ivanehreshi.interfaces;
 
 public interface GraphicManipulator {
-	public void update(double deltaTime);
+	public void update(long deltaTime);
 }
